refactor: share G<=h inequality normalization via an InequalityConstrained mixin
Both CLA (G w <= h exposure caps) and Lasso (G beta <= h) had independently
defined `g_matrix`, `h_vector`, and `event_dimension` -- the optional-inequality
input normalization and the n+p path-length coordinate count. `event_dimension`
was byte-identical; `g_matrix`/`h_vector` differed only in that the CLA cast g/h
to float64 and the LASSO did not.

Factor the three into an `InequalityConstrained` mixin in pathtracer.py (next to
the ParametricProblem protocol both problem types implement) and inherit it from
both classes. Each class keeps its own problem-specific `dimension`.

Side effect, and a small correctness fix: the LASSO now normalizes g/h to
float64 through the shared helper, so an integer-typed constraint array behaves
the same as it already did for the CLA.

No behaviour change otherwise: same paths, 257 tests pass, 100% coverage, ty +
mypy --strict clean.
